@@ -31,12 +31,12 @@
 		"@sveltejs/kit": "next",
 		"@types/geojson": "^7946.0.8",
 		"@types/matter-js": "^0.17.7",
-		"@types/node": "^17.0.30",
+		"@types/node": "^17.0.31",
 		"@types/overlayscrollbars": "^1.12.1",
 		"@types/svg-parser": "^2.0.3",
 		"@typescript-eslint/eslint-plugin": "^5.21.0",
 		"@typescript-eslint/parser": "^5.21.0",
-		"autoprefixer": "^10.4.5",
+		"autoprefixer": "^10.4.6",
 		"browserslist": "^4.20.3",
 		"cssnano": "^5.1.7",
 		"eslint": "^7.32.0",
@@ -49,7 +49,7 @@
 		"postcss-url": "^10.1.3",
 		"prettier": "^2.6.2",
 		"prettier-plugin-svelte": "^2.7.0",
-		"svelte": "^3.47.0",
+		"svelte": "^3.48.0",
 		"svelte-check": "^2.7.0",
 		"svelte-inline-svg": "^1.1.3",
 		"svelte-preprocess": "^4.10.6",
