@@ -4,32 +4,27 @@
 	import Legend from '$components/legend/Legend.svelte';
 	import LegendItem from '$components/legend/LegendItem.svelte';
 	import Loading from '$components/Loading.svelte';
+	import { LngLat, type LngLatLike } from 'maplibre-gl';
+	import { getContext } from 'svelte';
 
 	export let src: string;
 	export let highlightKey = null;
-	export let centered: boolean = false;
+	export let lnglat: LngLat | LngLatLike = undefined;
 
-	let max;
-	let center;
-	let visible = false;
-	function show() {
-		visible = true;
-	}
-	
-	async function fetchData() {
-		const r = await fetch(base + src);
-		if (r.ok) {
-			const data: BarChartData = await r.json();
-			max = Math.max(...data.columns.map(column => column.rows.reduce((sum, curr) => sum + curr, 0)));
-			center = Math.max(...data.columns.map(col => col.rows[0]));
-			return data;
-		}
-		else {
-			throw new Error(`Le chargement du fichier de données ${src} a encontré une erreur`);
-		}
-	}
+	const mapCtx = getContext('figuremap');
 
-	let fetchPromise = fetchData();
+	// async function fetchData() {
+	// 	const r = await fetch(base + src);
+	// 	if (r.ok) {
+	// 		const data: BarChartData = await r.json();
+	// 		max = Math.max(...data.columns.map(column => column.rows.reduce((sum, curr) => sum + curr, 0)));
+	// 		center = Math.max(...data.columns.map(col => col.rows[0]));
+	// 		return data;
+	// 	}
+	// 	else {
+	// 		throw new Error(`Le chargement du fichier de données ${src} a encontré une erreur`);
+	// 	}
+	// }
 
 	function setKey(e: MouseEvent) {
 		highlightKey = (e.target as HTMLElement).getAttribute('key');
@@ -45,5 +40,4 @@
 </svg>
 
 <style>
-
 </style>
