@@ -10,6 +10,7 @@
 	export let color: string = 'var(--dark1)';
 	export let colorHighlight: string = color;
 	export let label: string | number = null;
+	export let opacity: number = 1;
 	/* Fill color of the svg shape */
 	export let fill = 'none';
 	export let fillHighlight = fill;
@@ -19,6 +20,7 @@
 	export let strokeWidth = 1;
 	export let strokeWidthHighlight = strokeWidth;
 	export let strokeType: SymbolStrokeType = 'solid';
+	export let strokeDashArray: string = undefined;
 	/* Interactivity */
 	export let highlight = false;
 
@@ -26,19 +28,19 @@
 	const padding = 5;
 	const correctedPadding = padding * 1.5;
 	const textSize = size / 2.2;
-	const dashArray = strokeType === 'dashed' ? `${3 * strokeWidth}, ${1.5 * strokeWidth}` : strokeType === 'dotted' ? `0, ${2 * strokeWidth}` : null;
+	const dashArrayPreset = strokeType === 'dashed' ? `${3 * strokeWidth}, ${1.5 * strokeWidth}` : strokeType === 'dotted' ? `0, ${2 * strokeWidth}` : null;
 	let style;
 	$: style = {
 		'fill': highlight ? fillHighlight : fill,
 		'stroke': highlight ? strokeHighlight : stroke,
 		'stroke-width': highlight ? strokeWidthHighlight : strokeWidth,
 		'stroke-linecap': shape === 'dashed' ? 'butt' : ('round' as any),
-		'stroke-dasharray': dashArray || ''
+		'stroke-dasharray': strokeDashArray || dashArrayPreset || ''
 	};
 </script>
 
 {#if !src}
-	<svg width={size} height={size} viewBox="0 0 {size} {size}" {...$$restProps}>
+	<svg width={size} height={size} viewBox="0 0 {size} {size}" {...$$restProps} style:opacity>
 		{#if shape === 'square'}
 			<rect rx={size / 20} ry={size / 20} x={correctedPadding} y={correctedPadding} width={size - 2 * correctedPadding} height={size - 2 * correctedPadding} {...style} class:highlight />
 		{:else if shape === 'circle'}
@@ -46,7 +48,7 @@
 		{:else if shape === 'triangle'}
 			<polygon points="{size / 2} {padding}, {padding} {size - correctedPadding}, {size - padding} {size - correctedPadding}" {...style} class:highlight />
 		{:else if ['line', 'dotted', 'dashed'].indexOf(shape) > -1}
-			<path d="M{0},{(2 * size) / 3} Q{size / 4},{padding} {size / 2},{size / 2} T{size},{size / 3}" {...style} class:highlight />
+			<path d="M{padding},{size / 2} Q{size / 4},{size / 3} {size / 2},{size / 2} T{size - padding},{size / 2}" vector-effect="non-scaling-stroke" {...style} class:highlight />
 		{:else if Boolean(src)}
 			<!-- To implement! -->
 			<text>Source: {src}</text>
@@ -66,7 +68,7 @@
 		position: relative;
 		user-select: none;
 		width: 100%;
-		height: 100%;
+		height: auto;
 		overflow: visible;
 		margin: 0;
 		padding: 0;
