@@ -24,7 +24,10 @@
 			<ListItem><strong>Densité de population</strong> (ex.&nbsp;: nombre de personne au km<sup>2</sup>);</ListItem>
 			<ListItem><strong>Densité de construction</strong> (ex.&nbsp;: nombre de logements/hectare).</ListItem>
 		</List>
-		<p>Il est à noter que lorsqu’on parle de haute densité cela n’est pas synonyme d’immeuble de grande hauteur. Il est possible de construire des quartiers à haute densité, mais de faible hauteur.</p>
+		<p>
+			Il est à noter que lorsqu’on parle de haute densité cela n’est pas synonyme d’immeuble de grande hauteur. Il est possible de construire des quartiers à haute densité, mais de faible
+			hauteur.
+		</p>
 		<p>Le concept de densité fait aussi référence à des appréciations <strong>qualitatives</strong>, il est alors nécessaire de se pencher sur&nbsp;:</p>
 		<List>
 			<ListItem>Ses <strong>relations avec les qualités formelles</strong> générées à l’échelle du paysage&nbsp;urbain;</ListItem>
@@ -39,14 +42,22 @@
 		</Bodies>
 	</Concept>
 	<Interest heading="L’intérêt de nouveaux pôles de mixité et de densification à Saint-Constant">
-		<p>Saint-Constant est une ville au caractère familial, visible, entre autres, par la prédominance des habitations unifamiliales pavillonnaires. Cependant les études préalables révèlent&nbsp;:</p>
+		<p>
+			Saint-Constant est une ville au caractère familial, visible, entre autres, par la prédominance des habitations unifamiliales pavillonnaires. Cependant les études préalables révèlent&nbsp;:
+		</p>
 		<List>
 			<ListItem>La présence d’une diversification de la composition démographique qu’il faut prendre en compte dans les aménagements futurs</ListItem>
 			<ListItem>Un besoin de développer une diversité fonctionnelle du territoire pour favoriser une plus grande autonomie du milieu</ListItem>
-			<ListItem>La présence de trois aires T.O.D. (<i>Transit Oriented Developement</i>) soumises à̀ des paramètres de densification selon la réglementation du Plan métropolitain d’aménagement et de développement</ListItem>
-			<ListItem>Une pression à la densification qui s’exerce dans ce contexte autant en milieu construit (ex.&nbsp;: sur la rue Saint-Pierre), que dans les zones de requalification (ex.&nbsp;: la Route 132) ou dans les milieux à bâtir (ex.&nbsp;: La gare Sainte-Catherine)</ListItem>
+			<ListItem
+				>La présence de trois aires T.O.D. (<i>Transit Oriented Developement</i>) soumises à̀ des paramètres de densification selon la réglementation du Plan métropolitain d’aménagement et de
+				développement</ListItem
+			>
+			<ListItem
+				>Une pression à la densification qui s’exerce dans ce contexte autant en milieu construit (ex.&nbsp;: sur la rue Saint-Pierre), que dans les zones de requalification (ex.&nbsp;: la
+				Route 132) ou dans les milieux à bâtir (ex.&nbsp;: La gare Sainte-Catherine)</ListItem
+			>
 		</List>
-		<p><strong>Dans un tel contexte, il est donc primordial de penser des modèles de densification qui s’insèrent harmonieusement dans le paysage urbain de Saint-Constant.</strong></p>	
+		<p><strong>Dans un tel contexte, il est donc primordial de penser des modèles de densification qui s’insèrent harmonieusement dans le paysage urbain de Saint-Constant.</strong></p>
 	</Interest>
 	<Preferences>
 		<p>Issus de la démarche de caractérisation des paysages, les neuf secteurs de densification identifiés sont&nbsp;:</p>
@@ -69,7 +80,10 @@
 			<ListItem>La composition architecturale et urbaine des bâtiments;</ListItem>
 			<ListItem>La localisation des projets (selon les neuf secteurs de densification identifiés).</ListItem>
 		</List>
-		<p>Pour ce faire, six exemples de projets de densification, représentant une variété de typologies de bâtiments, ont été exposés aux participants. Ils se différenciaient par les caractéristiques suivantes&nbsp;:</p>
+		<p>
+			Pour ce faire, six exemples de projets de densification, représentant une variété de typologies de bâtiments, ont été exposés aux participants. Ils se différenciaient par les
+			caractéristiques suivantes&nbsp;:
+		</p>
 		<List>
 			<ListItem>La hauteur du bâtiment en étages;</ListItem>
 			<ListItem>La composition volumétrique;</ListItem>
@@ -81,9 +95,7 @@
 		<!-- figures des exemples -->
 		<h4>Résultats des préférences énoncées par les participants du rendez-vous citoyen</h4>
 		<FigureMap>
-			<FigureMarker label="1">
-				<!-- Piechart here -->
-			</FigureMarker>
+			<!-- Piechart here -->
 			<Legend>
 				<LegendItem label="1">Smart Centres Saint-Constant</LegendItem>
 				<LegendItem label="2">Voie de la Desserte</LegendItem>
@@ -97,22 +109,31 @@
 			</Legend>
 			<List>
 				<ListItem>Les immeubles de plus grandes hauteurs (ex.: A et B) présentent un faible intérêt, n’appartenant pas à l’identité de Saint-Constant selon les participants</ListItem>
-				<ListItem>Les immeubles de moyenne densité  (ex.:  C et D) présentent un certain intérêt. L’exemple C est préféré dans les secteurs en transformation (montée Saint-Régis, rue Bélanger et la rue Marois). L’exemple D est privilégié en bordure de la rue Saint-Régis et à l’intersection du chemin Sainte-Catherine et du boulevard Montchamp</ListItem>
-				<ListItem>Les immeubles de plus faible densité  (ex.: E et F) sont fortement appréciés. L’exemple E séduit pour son apparence abordable, il est préféré pour les sites de la gare Saint-Constant, la montée Saint-Régis et le site de la Chandellerie. L’exemple F est affectionné pour son caractère moderne, il est proposé pour la montée de Lasaline, la rue Saint-Pierre et le chemin Sainte-Catherine.</ListItem>
+				<ListItem
+					>Les immeubles de moyenne densité (ex.: C et D) présentent un certain intérêt. L’exemple C est préféré dans les secteurs en transformation (montée Saint-Régis, rue Bélanger et la
+					rue Marois). L’exemple D est privilégié en bordure de la rue Saint-Régis et à l’intersection du chemin Sainte-Catherine et du boulevard Montchamp</ListItem
+				>
+				<ListItem
+					>Les immeubles de plus faible densité (ex.: E et F) sont fortement appréciés. L’exemple E séduit pour son apparence abordable, il est préféré pour les sites de la gare
+					Saint-Constant, la montée Saint-Régis et le site de la Chandellerie. L’exemple F est affectionné pour son caractère moderne, il est proposé pour la montée de Lasaline, la rue
+					Saint-Pierre et le chemin Sainte-Catherine.</ListItem
+				>
 			</List>
 		</FigureMap>
-		<p>La hauteur des bâtiments demeure un élément d’appréciation négatif des projets de densification. Aussi, la présence de la végétation ainsi que de l’activité commerciale contribuent à apaiser la perception négative des projets. Nonobstant ces commentaires généraux, il est à noter que certaines tables de discussion ont été en défaveur de tous les exemples présentés.</p>
+		<p>
+			La hauteur des bâtiments demeure un élément d’appréciation négatif des projets de densification. Aussi, la présence de la végétation ainsi que de l’activité commerciale contribuent à
+			apaiser la perception négative des projets. Nonobstant ces commentaires généraux, il est à noter que certaines tables de discussion ont été en défaveur de tous les exemples présentés.
+		</p>
 	</Preferences>
 	<Propositions>
-		<p>À la suite des analyses précédemment présentées, un plan stratégique de densification pour les secteurs urbains identifiés a été établi. Chacun des secteurs sont à vocation mixte pour permettre de développer une vie de quartier singulière tout en respectant le caractère paysager déjà présent.</p>
-		<Proposition
-			label="Secteur A"
-			title="Le site de la chandellerie"
-			key="a"
-		>
+		<p>
+			À la suite des analyses précédemment présentées, un plan stratégique de densification pour les secteurs urbains identifiés a été établi. Chacun des secteurs sont à vocation mixte pour
+			permettre de développer une vie de quartier singulière tout en respectant le caractère paysager déjà présent.
+		</p>
+		<Proposition label="Secteur A" title="Le site de la chandellerie" key="a">
 			<h4>Contexte, forces et faiblesses</h4>
 			<Table cols="3" style="max-width: var(--width-md)">
-				<TableCell type="spacer"></TableCell>
+				<TableCell type="spacer" />
 				<TableCell type="heading" align="center">Forces</TableCell>
 				<TableCell type="heading" align="center">Faiblesses</TableCell>
 				<TableCell type="heading" orientation="vertical">Noyau villageois</TableCell>
@@ -142,25 +163,31 @@
 				<TableCell type="heading" align="center">Densité recommandée</TableCell>
 				<TableCell align="center"><strong>A</strong>, <strong>B</strong></TableCell>
 				<TableCell align="center">Rue Saint-Pierre et Montée Saint-Régis</TableCell>
-				<TableCell align="center">1 à 2 étages ou<br>5 à 8 mètres</TableCell>
+				<TableCell align="center">1 à 2 étages ou<br />5 à 8 mètres</TableCell>
 				<TableCell align="center">Commercial et résidentiel</TableCell>
 				<TableCell align="center">Projets immobiliers à basse densité</TableCell>
 			</Table>
 			<h4>Ce que le projet propose</h4>
 			<List>
-				<ListItem>Construire un ensemble poreux grâce à la disposition des bâtiments de logements perpendiculaires à la rue. Cela permet de créer des liens piétons et visuels vers la rivière Saint-Pierre et de mettre en valeur ce milieu naturel. De plus, cela réduit l’impact de façades linéaires sur la rue, au risque d’être trop monumentales, ce qui engendrerait une rupture avec le contexte;</ListItem>
-				<ListItem>Accorder les bâtiments avec le cadre existant. Ils font référence aux constructions localisées dans le milieu agricole (toits à versants, matériau bois favorisé). Pour atténuer l’effet de hauteur, des terrasses sont proposées en toiture permettant un recul de la façade;</ListItem>
-				<ListItem>Aménager une place publique et des commerces de proximités en rez-de-chaussée des bâtiments qui donne sur celle-ci. Ces services apporteront de l’animation au site et favoriseront le développement d’une vie de quartier.</ListItem>
+				<ListItem
+					>Construire un ensemble poreux grâce à la disposition des bâtiments de logements perpendiculaires à la rue. Cela permet de créer des liens piétons et visuels vers la rivière
+					Saint-Pierre et de mettre en valeur ce milieu naturel. De plus, cela réduit l’impact de façades linéaires sur la rue, au risque d’être trop monumentales, ce qui engendrerait une
+					rupture avec le contexte;</ListItem
+				>
+				<ListItem
+					>Accorder les bâtiments avec le cadre existant. Ils font référence aux constructions localisées dans le milieu agricole (toits à versants, matériau bois favorisé). Pour atténuer
+					l’effet de hauteur, des terrasses sont proposées en toiture permettant un recul de la façade;</ListItem
+				>
+				<ListItem
+					>Aménager une place publique et des commerces de proximités en rez-de-chaussée des bâtiments qui donne sur celle-ci. Ces services apporteront de l’animation au site et favoriseront
+					le développement d’une vie de quartier.</ListItem
+				>
 			</List>
 		</Proposition>
-		<Proposition
-			label="Secteur B"
-			title="La montée Saint-Régis"
-			key="b"
-		>
+		<Proposition label="Secteur B" title="La montée Saint-Régis" key="b">
 			<h4>Contexte, forces et faiblesses</h4>
 			<Table cols="3" style="max-width: var(--width-md)">
-				<TableCell type="spacer"></TableCell>
+				<TableCell type="spacer" />
 				<TableCell type="heading" align="center">Forces</TableCell>
 				<TableCell type="heading" align="center">Faiblesses</TableCell>
 				<TableCell type="heading" orientation="vertical">Noyau villageois</TableCell>
@@ -189,25 +216,24 @@
 				<TableCell type="heading" align="center">Densité recommandée</TableCell>
 				<TableCell align="center"><strong>A</strong>, <strong>B</strong>, <strong>C</strong></TableCell>
 				<TableCell align="center">Montée Saint-Régis</TableCell>
-				<TableCell align="center">1 à 4 étages ou<br>5 à 16 mètres</TableCell>
+				<TableCell align="center">1 à 4 étages ou<br />5 à 16 mètres</TableCell>
 				<TableCell align="center">Commercial et résidentiel</TableCell>
 				<TableCell align="center">Projets moyenne densité (similaire à l'existant)</TableCell>
 			</Table>
 			<h4>Ce que le projet propose</h4>
 			<List>
-				<ListItem>Intégrer des axes piétons végétalisés, sécuritaires afin de relier l’école, les commerces, les résidences et le terrain de jeux présents. Ces axes piétons desserviraient les entrées des bâtiments familiaux projetés. Les façades côté rue accueilleraient les commerces de proximité;</ListItem>
+				<ListItem
+					>Intégrer des axes piétons végétalisés, sécuritaires afin de relier l’école, les commerces, les résidences et le terrain de jeux présents. Ces axes piétons desserviraient les
+					entrées des bâtiments familiaux projetés. Les façades côté rue accueilleraient les commerces de proximité;</ListItem
+				>
 				<ListItem>Implanter des bâtiments de logements reprenant les principes typologiques et de gabarits développés dans le secteur A;</ListItem>
 				<ListItem>Aménager des aires de stationnement subdivisées en plusieurs lots réparties sur le site afin de réduire leur perception sur le plan visuel.</ListItem>
 			</List>
 		</Proposition>
-		<Proposition
-			label="Secteur C"
-			title="L’intersection de la rue Sainte-Catherine et de la montée Saint-Régis"
-			key="c"
-		>
+		<Proposition label="Secteur C" title="L’intersection de la rue Sainte-Catherine et de la montée Saint-Régis" key="c">
 			<h4>Contexte, forces et faiblesses</h4>
 			<Table cols="3" style="max-width: var(--width-md)">
-				<TableCell type="spacer"></TableCell>
+				<TableCell type="spacer" />
 				<TableCell type="heading" align="center">Forces</TableCell>
 				<TableCell type="heading" align="center">Faiblesses</TableCell>
 				<TableCell type="heading" orientation="vertical">Noyau villageois</TableCell>
@@ -242,19 +268,22 @@
 			</Table>
 			<h4>Ce que le projet propose</h4>
 			<List>
-				<ListItem>Greffer des serres aux façades des bâtiments résidentiels. Ce langage formel et fonctionnel, contribue à unifier l’environnement urbain au milieu agricole. Les serres individuelles permettent d’agrandir le logement en offrant un balcon quatre saisons. Des serres collectives sont aussi proposées pour servir d’aires communes à un même immeuble. Finalement, des serres commerciales peuvent également s’implanter en toiture pour profiter de la chaleur des logements et pratiquer l’agriculture urbaine;</ListItem>
-				<ListItem>Implanter des commerces et des espaces publics en rez-de-chaussée des bâtiments sur le chemin Sainte-Catherine et la montée Saint-Régis, pour pallier le manque de commerces de proximités dans ce secteur. Cela répondra aux concepts d’aires T.O.D préconisant des aménagements denses et diversifiés;</ListItem>
+				<ListItem
+					>Greffer des serres aux façades des bâtiments résidentiels. Ce langage formel et fonctionnel, contribue à unifier l’environnement urbain au milieu agricole. Les serres
+					individuelles permettent d’agrandir le logement en offrant un balcon quatre saisons. Des serres collectives sont aussi proposées pour servir d’aires communes à un même immeuble.
+					Finalement, des serres commerciales peuvent également s’implanter en toiture pour profiter de la chaleur des logements et pratiquer l’agriculture urbaine;</ListItem
+				>
+				<ListItem
+					>Implanter des commerces et des espaces publics en rez-de-chaussée des bâtiments sur le chemin Sainte-Catherine et la montée Saint-Régis, pour pallier le manque de commerces de
+					proximités dans ce secteur. Cela répondra aux concepts d’aires T.O.D préconisant des aménagements denses et diversifiés;</ListItem
+				>
 				<ListItem>Faire des connexions piétonnes entre la gare Sainte-Catherine et l’agroparc.</ListItem>
 			</List>
 		</Proposition>
-		<Proposition
-			label="Secteur D"
-			title="Le site entre la gare Sainte-Catherine et le lac des fées"
-			key="d"
-		>
+		<Proposition label="Secteur D" title="Le site entre la gare Sainte-Catherine et le lac des fées" key="d">
 			<h4>Contexte, forces et faiblesses</h4>
 			<Table cols="3" style="max-width: var(--width-md)">
-				<TableCell type="spacer"></TableCell>
+				<TableCell type="spacer" />
 				<TableCell type="heading" align="center">Forces</TableCell>
 				<TableCell type="heading" align="center">Faiblesses</TableCell>
 				<TableCell type="heading" orientation="vertical">Noyau villageois</TableCell>
@@ -292,19 +321,21 @@
 			</Table>
 			<h4>Ce que le projet propose</h4>
 			<List>
-				<ListItem>Implanter des services de proximité au rez-de-chaussée des bâtiments face au chemin Sainte-Catherine. L’aménagement d’une place publique permet l’installation de terrasses selon la nature des commerces qui s’y implantent. Cette place est reliée à un réseau piéton qui mène à la gare Saint-Catherine;</ListItem>
-				<ListItem>Préserver et mettre en valeur les milieux naturels par l’aménagement de connexions visuelles et physiques entre le lac des Fées et l’axe des voies ferrées. De plus, l’aménagement de la majorité des places de stationnements seraient en souterrain;</ListItem>
+				<ListItem
+					>Implanter des services de proximité au rez-de-chaussée des bâtiments face au chemin Sainte-Catherine. L’aménagement d’une place publique permet l’installation de terrasses selon
+					la nature des commerces qui s’y implantent. Cette place est reliée à un réseau piéton qui mène à la gare Saint-Catherine;</ListItem
+				>
+				<ListItem
+					>Préserver et mettre en valeur les milieux naturels par l’aménagement de connexions visuelles et physiques entre le lac des Fées et l’axe des voies ferrées. De plus, l’aménagement
+					de la majorité des places de stationnements seraient en souterrain;</ListItem
+				>
 				<ListItem>Conserver un recul important vis-à-vis de l’emprise ferroviaire avec les bâtiments de logements pour éviter les nuisances engendrées (bruits, vibrations etc.).</ListItem>
 			</List>
 		</Proposition>
-		<Proposition
-			label="Secteur E"
-			title="La rue Sainte-Catherine à l’intersection de la rue Meunier"
-			key="e"
-		>
+		<Proposition label="Secteur E" title="La rue Sainte-Catherine à l’intersection de la rue Meunier" key="e">
 			<h4>Contexte, forces et faiblesses</h4>
 			<Table cols="3" style="max-width: var(--width-md)">
-				<TableCell type="spacer"></TableCell>
+				<TableCell type="spacer" />
 				<TableCell type="heading" align="center">Forces</TableCell>
 				<TableCell type="heading" align="center">Faiblesses</TableCell>
 				<TableCell type="heading" orientation="vertical">Noyau villageois</TableCell>
@@ -332,24 +363,23 @@
 				<TableCell type="heading" align="center">Densité recommandée</TableCell>
 				<TableCell align="center"><strong>A</strong>, <strong>B</strong></TableCell>
 				<TableCell align="center">Rue Sainte-Catherine / Meunier</TableCell>
-				<TableCell align="center">1 à 2 étages ou<br>5 à 10 mètres</TableCell>
+				<TableCell align="center">1 à 2 étages ou<br />5 à 10 mètres</TableCell>
 				<TableCell align="center">Commercial et résidentiel</TableCell>
 				<TableCell align="center">Projets immobiliers de basse à moyenne densité</TableCell>
 			</Table>
 			<h4>Ce que le projet propose</h4>
 			<List>
-				<ListItem>Favoriser l’interaction directe des locaux commerciaux avec l’espace public. La reconstruction des bâtiments commerciaux en front de rue, le déplacement des stationnements en arrière-cours, l’ajout de placettes publiques et l’aménagement du trottoir bordé d’arbres, permettraient cette interaction et plus de convivialité;</ListItem>
+				<ListItem
+					>Favoriser l’interaction directe des locaux commerciaux avec l’espace public. La reconstruction des bâtiments commerciaux en front de rue, le déplacement des stationnements en
+					arrière-cours, l’ajout de placettes publiques et l’aménagement du trottoir bordé d’arbres, permettraient cette interaction et plus de convivialité;</ListItem
+				>
 				<ListItem>Ajouter des unités de logements au-dessus des commerces.</ListItem>
 			</List>
 		</Proposition>
-		<Proposition
-			label="Secteur F"
-			title="Requalification de la 132"
-			key="f"
-		>
+		<Proposition label="Secteur F" title="Requalification de la 132" key="f">
 			<h4>Contexte, forces et faiblesses</h4>
 			<Table cols="3" style="max-width: var(--width-md)">
-				<TableCell type="spacer"></TableCell>
+				<TableCell type="spacer" />
 				<TableCell type="heading" align="center">Forces</TableCell>
 				<TableCell type="heading" align="center">Faiblesses</TableCell>
 				<TableCell type="heading" orientation="vertical">Noyau villageois</TableCell>
@@ -369,7 +399,10 @@
 				</TableCell>
 			</Table>
 			<h4>Informations relatives au site projeté</h4>
-			<p>Le projet vise à illustrer le potentiel de densification tout en examinant les potentiels de mise en valeur des parcours d’entrées de ville tirant profit de l’identité de Saint-Constant.</p>
+			<p>
+				Le projet vise à illustrer le potentiel de densification tout en examinant les potentiels de mise en valeur des parcours d’entrées de ville tirant profit de l’identité de
+				Saint-Constant.
+			</p>
 			<Table cols="5" style="max-width: var(--width-md)">
 				<TableCell type="heading" align="center">Exemples</TableCell>
 				<TableCell type="heading" align="center">Implantation</TableCell>
@@ -384,18 +417,23 @@
 			</Table>
 			<h4>Ce que le projet propose</h4>
 			<List>
-				<ListItem>Créer un axe commercial poreux et convivial parallèle à la route 132. Le projet suggère d’aménager une rue de desserte locale, parallèle à l’artère pour favoriser un environnement à échelle humaine privilégiant la circulation douce. Cette rue serait encadrée de commerces aux rez-de-chaussée des bâtiments et composée de terrasses et d’installations publiques. Finalement, l’axe se connecterait au nouveau pôle récréatif régional (complexe aquatique, aréna) et offrirait un axe de mobilité active vert et convivial;</ListItem>
-				<ListItem>Adapter la hauteur des bâtiments de logements en fonction du contexte actuel. Ainsi, du côté sud du boulevard, secteur moins dense, des bâtiments de faible hauteur seront favorisés. Tandis que du côté nord des bâtiments de plus grande hauteur sont suggérés. Pour éviter un linéaire de façade trop important, qui donnerait l’effet d’un mur et d’une ville très dense, il est proposé de maintenir des espaces verts transversaux à la route 132. Il est conseillé de favoriser des bâtiments de faibles hauteurs à leurs pourtours pour diminuer les ombres portées.</ListItem>
+				<ListItem
+					>Créer un axe commercial poreux et convivial parallèle à la route 132. Le projet suggère d’aménager une rue de desserte locale, parallèle à l’artère pour favoriser un environnement
+					à échelle humaine privilégiant la circulation douce. Cette rue serait encadrée de commerces aux rez-de-chaussée des bâtiments et composée de terrasses et d’installations publiques.
+					Finalement, l’axe se connecterait au nouveau pôle récréatif régional (complexe aquatique, aréna) et offrirait un axe de mobilité active vert et convivial;</ListItem
+				>
+				<ListItem
+					>Adapter la hauteur des bâtiments de logements en fonction du contexte actuel. Ainsi, du côté sud du boulevard, secteur moins dense, des bâtiments de faible hauteur seront
+					favorisés. Tandis que du côté nord des bâtiments de plus grande hauteur sont suggérés. Pour éviter un linéaire de façade trop important, qui donnerait l’effet d’un mur et d’une
+					ville très dense, il est proposé de maintenir des espaces verts transversaux à la route 132. Il est conseillé de favoriser des bâtiments de faibles hauteurs à leurs pourtours pour
+					diminuer les ombres portées.</ListItem
+				>
 			</List>
 		</Proposition>
-		<Proposition
-			label="Secteur G"
-			title="la rue Saint-Pierre"
-			key="g"
-		>
+		<Proposition label="Secteur G" title="la rue Saint-Pierre" key="g">
 			<h4>Contexte, forces et faiblesses</h4>
 			<Table cols="3" style="max-width: var(--width-md)">
-				<TableCell type="spacer"></TableCell>
+				<TableCell type="spacer" />
 				<TableCell type="heading" align="center">Forces</TableCell>
 				<TableCell type="heading" align="center">Faiblesses</TableCell>
 				<TableCell type="heading" orientation="vertical">Noyau villageois</TableCell>
@@ -415,7 +453,10 @@
 				</TableCell>
 			</Table>
 			<h4>Informations relatives au site projeté</h4>
-			<p>Le projet vise à illustrer le potentiel de densification tout en examinant les potentiels de mise en valeur des parcours d’entrées de ville tirant profit de l’identité de Saint-Constant.</p>
+			<p>
+				Le projet vise à illustrer le potentiel de densification tout en examinant les potentiels de mise en valeur des parcours d’entrées de ville tirant profit de l’identité de
+				Saint-Constant.
+			</p>
 			<Table cols="5" style="max-width: var(--width-md)">
 				<TableCell type="heading" align="center">Exemples</TableCell>
 				<TableCell type="heading" align="center">Implantation</TableCell>
@@ -424,26 +465,29 @@
 				<TableCell type="heading" align="center">Densité recommandée</TableCell>
 				<TableCell align="center"><strong>A</strong>, <strong>B</strong>, <strong>C</strong></TableCell>
 				<TableCell align="center">Le long de la rue Saint-Pierre sur un axe commercial et résidentiel</TableCell>
-				<TableCell align="center">1 à 4 étages ou<br>5 à 16 mètres</TableCell>
+				<TableCell align="center">1 à 4 étages ou<br />5 à 16 mètres</TableCell>
 				<TableCell align="center">Commercial et résidentiel</TableCell>
 				<TableCell align="center">Projets immobiliers de basse à moyenne densité</TableCell>
 			</Table>
 			<h4>Ce que le projet propose</h4>
 			<List>
-				<ListItem>Réduire les contrastes d’échelle et de hauteur visibles entre les nouvelles et les anciennes constructions. À cet effet, il est proposé d’insérer les bâtiments perpendiculairement pour orienter les plus petites façades vers la rue Saint-Pierre. Des retraits des derniers étages sont recommandés pour atténuer la sensation de hauteur et donc de densité;</ListItem>
+				<ListItem
+					>Réduire les contrastes d’échelle et de hauteur visibles entre les nouvelles et les anciennes constructions. À cet effet, il est proposé d’insérer les bâtiments perpendiculairement
+					pour orienter les plus petites façades vers la rue Saint-Pierre. Des retraits des derniers étages sont recommandés pour atténuer la sensation de hauteur et donc de densité;</ListItem
+				>
 				<ListItem>Insérer aux bâtiments des commerces aux rez-de-chaussée pour faire des milieux mixtes et fonctionnels;</ListItem>
-				<ListItem>Apporter une grande attention au verdissement de la rue Saint-Pierre, actuellement très minérale. Les surfaces végétales des lots privés doivent être maintenues, l’aménagement d’une portion d’aires de stationnements souterraines est suggéré. Les aires en surfaces devraient être maintenues à l’arrière des lots pour permettre un verdissement des façades avant;</ListItem>
+				<ListItem
+					>Apporter une grande attention au verdissement de la rue Saint-Pierre, actuellement très minérale. Les surfaces végétales des lots privés doivent être maintenues, l’aménagement
+					d’une portion d’aires de stationnements souterraines est suggéré. Les aires en surfaces devraient être maintenues à l’arrière des lots pour permettre un verdissement des façades
+					avant;</ListItem
+				>
 				<ListItem>Aménager la rue pour favoriser la marche comme des allées d’arbres, du mobilier urbain etc.</ListItem>
 			</List>
 		</Proposition>
-		<Proposition
-			label="Secteur H"
-			title="Le site de la gare Saint-Constant"
-			key="h"
-		>
+		<Proposition label="Secteur H" title="Le site de la gare Saint-Constant" key="h">
 			<h4>Contexte, forces et faiblesses</h4>
 			<Table cols="3" style="max-width: var(--width-md)">
-				<TableCell type="spacer"></TableCell>
+				<TableCell type="spacer" />
 				<TableCell type="heading" align="center">Forces</TableCell>
 				<TableCell type="heading" align="center">Faiblesses</TableCell>
 				<TableCell type="heading" orientation="vertical">Noyau villageois</TableCell>
@@ -465,7 +509,10 @@
 				</TableCell>
 			</Table>
 			<h4>Informations relatives au site projeté</h4>
-			<p>Le projet vise à illustrer le potentiel de densification tout en examinant les potentiels de mise en valeur des parcours d’entrées de ville tirant profit de l’identité de Saint-Constant.</p>
+			<p>
+				Le projet vise à illustrer le potentiel de densification tout en examinant les potentiels de mise en valeur des parcours d’entrées de ville tirant profit de l’identité de
+				Saint-Constant.
+			</p>
 			<Table cols="5" style="max-width: var(--width-md)">
 				<TableCell type="heading" align="center">Exemples</TableCell>
 				<TableCell type="heading" align="center">Implantation</TableCell>
@@ -474,15 +521,26 @@
 				<TableCell type="heading" align="center">Densité recommandée</TableCell>
 				<TableCell align="center"><strong>A</strong>, <strong>B</strong>, <strong>C</strong></TableCell>
 				<TableCell align="center">Le long de la rue Saint-Pierre et d'un axe ferroviaire</TableCell>
-				<TableCell align="center">1 à 2 étages ou<br>5 à 10 mètres</TableCell>
+				<TableCell align="center">1 à 2 étages ou<br />5 à 10 mètres</TableCell>
 				<TableCell align="center">Commercial et résidentiel</TableCell>
 				<TableCell align="center">Projets de moyenne à haute densité avec ajout de stationnement</TableCell>
 			</Table>
 			<h4>Ce que le projet propose</h4>
 			<List>
-				<ListItem>Implanter une végétalisation massive des aires de stationnements afin d’améliorer la qualité de l’environnement en termes d’usages et de bilan écologique. La proposition d’aménagement suggère l’ajout de bandes de plantations et des noues paysagères pour récolter les eaux de ruissellement pluviales entre les rangs de stationnements. L’ajout d’arbres permettrait de générer de l’ombre, contribuant à réduire les effets d’îlots de chaleur. Des stratégies semblent également envisageables pour réduire le nombre de stationnements condensés dans un même vaste espace. Par exemple, des stationnements incitatifs pourraient être aménagés pour le soir et les fins de semaines;</ListItem>
-				<ListItem>Insérer des chemins piétons le long des commerces et entre les rangs de stationnements le long des bandes de plantations, reliés à la rue Saint-Pierre et la gare de Saint-Constant. Cette stratégie permettrait d’apporter une plus grande convivialité à ce site;</ListItem>
-				<ListItem>Ajouter un bâtiment aux fonctions mixtes avec des commerces en rez-de-chaussée (ex. : aires de “coworking”, cafés, atelier de vélo etc.) le long de la rue Pacifique. Son emplacement permettrait de séparer le milieu résidentiel des stationnements incitatifs.</ListItem>
+				<ListItem
+					>Implanter une végétalisation massive des aires de stationnements afin d’améliorer la qualité de l’environnement en termes d’usages et de bilan écologique. La proposition
+					d’aménagement suggère l’ajout de bandes de plantations et des noues paysagères pour récolter les eaux de ruissellement pluviales entre les rangs de stationnements. L’ajout d’arbres
+					permettrait de générer de l’ombre, contribuant à réduire les effets d’îlots de chaleur. Des stratégies semblent également envisageables pour réduire le nombre de stationnements
+					condensés dans un même vaste espace. Par exemple, des stationnements incitatifs pourraient être aménagés pour le soir et les fins de semaines;</ListItem
+				>
+				<ListItem
+					>Insérer des chemins piétons le long des commerces et entre les rangs de stationnements le long des bandes de plantations, reliés à la rue Saint-Pierre et la gare de
+					Saint-Constant. Cette stratégie permettrait d’apporter une plus grande convivialité à ce site;</ListItem
+				>
+				<ListItem
+					>Ajouter un bâtiment aux fonctions mixtes avec des commerces en rez-de-chaussée (ex. : aires de “coworking”, cafés, atelier de vélo etc.) le long de la rue Pacifique. Son
+					emplacement permettrait de séparer le milieu résidentiel des stationnements incitatifs.</ListItem
+				>
 			</List>
 		</Proposition>
 	</Propositions>
